Better typesetting
Changed the join point so that we join in the renderer (which makes more sense), and added an empty string at the start (which fixes indenting the descendant lines after the first by indenting the first line to match).

This simply shouldn't be done through string.join at this point, and I should switch to a string buffer system instead. That's a tomorrow me problem.